round-3/web03/src/notes/requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pycparser==2.22
PyJWT==2.8.0
sqlitedict==2.1.0
supervisor==4.2.1
uWSGI==2.0.24
Werkzeug==3.0.2
requests